Refactor and optimize code for script
