+++ JSchema/RelogicLabs/JSchema/Utilities/CommonUtilities.cs
@@ -1,8 +1,4 @@
-using RelogicLabs.JsonSchema.Exceptions;
-using RelogicLabs.JsonSchema.Types;
-using static RelogicLabs.JsonSchema.Message.MessageFormatter;
-
-namespace RelogicLabs.JsonSchema.Utilities;
+namespace RelogicLabs.JSchema.Utilities;
 
 internal static class CommonUtilities
 {
@@ -19,20 +15,5 @@ internal static class CommonUtilities
         return list.AsReadOnly();
     }
 
-    internal static JNode GetDerived(JNode target)
-    {
-        if(target is IDerived _derived) return _derived.Derived ?? target;
-        return target;
-    }
-
-    internal static IEnumerable<JProperty> RequireUniqueness(List<JProperty> list, string errorCode)
-    {
-        var group = list.GroupBy(p => p.Key).FirstOrDefault(g => g.Count() > 1);
-        if(group == default) return list;
-        JProperty property = group.First();
-        throw new DuplicatePropertyKeyException(FormatForJson(errorCode,
-            $"Multiple key with name '{property.Key}' found", property));
-    }
-
-    internal static T? TryGetLast<T>(List<T> list) => list.IsEmpty() ? default : list[^1];
+    internal static bool HasFlag(int flagSet, int flag) => (flagSet & flag) == flag;
 }
